move mod validation_function_params into its own file
